pyproject.toml
[tool.poetry]
name = "glucose-app"
version = "0.1.0"
description = ""
authors = ["Luciano Simeão"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.0.2"
django-filter = "^21.1"
markdown = "^3.3.6"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
python-decouple = "^3.6"
django-cors-headers = "^3.11.0"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
